--- src/Repository/Repository.php
@@ -1,15 +1,0 @@
-<?php
-namespace Application\Repository;
-
-
-class Repository
-{
-
-    protected $database;
-
-
-    public function __construct()
-    {
-        $this->database = new \PDO();
-    }
-}
