+++ src/Repository/UserRepository.php
@@ -2,30 +2,19 @@
 
 namespace Application\Repository;
 
-use Application\Application\Database\PDODatabase;
+use Application\Repository\AbstractRepository;
 
-class UserRepository
+class UserRepository extends AbstractRepository
 {
-    protected $database;
-
-    public function __construct()
+    protected function getTableName(): string
     {
-        try {
-            $this->database = new PDODatabase(
-                'mysql:host=127.0.0.1;dbname=myblogphp;charset=utf8',
-                'root',
-                '',
-            );
-        } catch (\Exception $e) {
-            dump($e);
-        }
+        return 'user';
     }
 
     public function findByEmail(string $email)
     {
-        $query = 'SELECT * FROM user WHERE email = :email';
-        $statement = $this->database->request($query, [':email' => $email]);;
-        return $statement->fetch;
+        $query = "SELECT * FROM {$this->getTableName()} WHERE email = :email";
+        $statement = $this->database->request($query, [':email' => $email]);
+        return $statement->fetch();
     }
-
 }
