+++ var/cache/twig/c9/c9b520347785c562c26e36058fdad4bddbca12a99e0911a0719988e18a5b2077.php
@@ -52,6 +52,7 @@ class __TwigTemplate_0d701f946a879aa126ddaf9d26a3f595f9c9cf049f9854193a4f17ccf38
 
 <body>
 
+
         <header class=\"sticky-top\">
                 <nav class=\"navbar navbar-dark navbar-expand-md bg-dark m-0 p-0\">
                     <div class=\" menunavigation container-fluid m-0 p-0\">
@@ -82,11 +83,13 @@ class __TwigTemplate_0d701f946a879aa126ddaf9d26a3f595f9c9cf049f9854193a4f17ccf38
                 </nav>
             </header>
 
+
 ";
-        // line 44
+        // line 46
         $this->displayBlock('body', $context, $blocks);
-        // line 47
+        // line 49
         echo "
+
 <footer>
         <div class=\"container-fluid bg-dark m-0 p-0\">
             <div id=\"row_footer\" class=\"row m-0 px-0 py-3\">
@@ -109,6 +112,7 @@ class __TwigTemplate_0d701f946a879aa126ddaf9d26a3f595f9c9cf049f9854193a4f17ccf38
             </div>
         </div>
     </footer>
+
 
         <!-- jQuery first, then Popper.js, then Bootstrap JS -->
         <script src=\"https://code.jquery.com/jquery-3.4.1.slim.min.js\" integrity=\"sha384-J6qa4849blE2+poT4WnyKhv5vZF5SrPo0iEjwBvKU7imGFAV0wwj1yYfoRSJoZ+n\" crossorigin=\"anonymous\"></script>
@@ -126,11 +130,11 @@ class __TwigTemplate_0d701f946a879aa126ddaf9d26a3f595f9c9cf049f9854193a4f17ccf38
         $macros = $this->macros;
     }
 
-    // line 44
+    // line 46
     public function block_body($context, array $blocks = [])
     {
         $macros = $this->macros;
-        // line 45
+        // line 47
         echo "
 ";
     }
@@ -142,7 +146,7 @@ class __TwigTemplate_0d701f946a879aa126ddaf9d26a3f595f9c9cf049f9854193a4f17ccf38
 
     public function getDebugInfo()
     {
-        return array (  134 => 45,  130 => 44,  124 => 6,  89 => 47,  87 => 44,  46 => 6,  39 => 1,);
+        return array (  138 => 47,  134 => 46,  128 => 6,  91 => 49,  89 => 46,  46 => 6,  39 => 1,);
     }
 
     public function getSourceContext()
@@ -159,6 +163,7 @@ class __TwigTemplate_0d701f946a879aa126ddaf9d26a3f595f9c9cf049f9854193a4f17ccf38
 </head>
 
 <body>
+
 
         <header class=\"sticky-top\">
                 <nav class=\"navbar navbar-dark navbar-expand-md bg-dark m-0 p-0\">
@@ -190,9 +195,11 @@ class __TwigTemplate_0d701f946a879aa126ddaf9d26a3f595f9c9cf049f9854193a4f17ccf38
                 </nav>
             </header>
 
+
 {% block body %}
 
 {% endblock %}
+
 
 <footer>
         <div class=\"container-fluid bg-dark m-0 p-0\">
@@ -216,6 +223,7 @@ class __TwigTemplate_0d701f946a879aa126ddaf9d26a3f595f9c9cf049f9854193a4f17ccf38
             </div>
         </div>
     </footer>
+
 
         <!-- jQuery first, then Popper.js, then Bootstrap JS -->
         <script src=\"https://code.jquery.com/jquery-3.4.1.slim.min.js\" integrity=\"sha384-J6qa4849blE2+poT4WnyKhv5vZF5SrPo0iEjwBvKU7imGFAV0wwj1yYfoRSJoZ+n\" crossorigin=\"anonymous\"></script>
